Initial commit of Raptor Invoice
Raptor Invoice is a fully working, good looking invoice application for
Windows. It is written in Delphi and targets the Norwegian market. It is
given away freely. You are however not allowed to re-package or re-brand
this, nor use the source code for personal gain.
This is a free product, but not without ownership.